+++ src/modules/UI/components/ControlPanel/ControlPanel.ui.js
@@ -2,53 +2,37 @@
 
 import * as React from 'react'
 import { View } from 'react-native'
-import { shallowEqual } from 'react-redux'
 
 import { selectWalletFromModal } from '../../../../actions/WalletActions.js'
 import { SceneWrapper } from '../../../../components/common/SceneWrapper.js'
 import { type Theme, cacheStyles, useTheme } from '../../../../components/services/ThemeContext'
 import Separator from '../../../../components/themed/Separator'
-import { RootState } from '../../../../types/reduxTypes.js'
-import type { GuiDenomination, GuiExchangeRates, GuiWallet } from '../../../../types/types.js'
-import { getCurrencyIcon } from '../../../../util/CurrencyInfoHelpers.js'
-import { useDispatch, useMemo, useSelector, useState } from '../../../../util/hooks'
-import { getDenomFromIsoCode } from '../../../../util/utils.js'
+import { type RootState } from '../../../../reducers/RootReducer.js'
+import { useDispatch, useSelector, useState } from '../../../../util/hooks'
+import { reduxShallowEqual } from '../../../../util/utils.js'
 import { logoutRequest } from '../../../Login/action'
-import { getDisplayDenominationFull } from '../../../Settings/selectors.js'
-import { getExchangeDenomination, getExchangeRate, getSelectedWallet } from '../../../UI/selectors.js'
-import PanelBody from './components/PanelBody'
-import PanelHeader from './components/PanelHeader'
+import AccountSwitcher from '../AccountSwitcher/AccountSwitcher.ui'
+import PanelCurrency from './components/PanelCurrency'
+import PanelDisable from './components/PanelDisable'
+import PanelList from './components/PanelList'
+import PanelLogo from './components/PanelLogo'
 
 export type StateProps = {
   selectedCurrencyCode: string,
-  selectedWalletId: string,
-  exchangeRates: GuiExchangeRates,
-  guiWallet: GuiWallet,
-  exchangeRate?: number,
-  primaryDisplayDenomination: GuiDenomination,
-  primaryExchangeDenomination: GuiDenomination
+  selectedWalletId: string
 }
 
-const selector = (state: RootState) => {
-  const guiWallet = getSelectedWallet(state)
-  const selectedCurrencyCode = state.ui.wallets.selectedCurrencyCode
-
-  return {
-    selectedWalletId: state.ui.wallets.selectedWalletId,
-    guiWallet,
-    selectedCurrencyCode,
-    exchangeRate: guiWallet ? getExchangeRate(state, selectedCurrencyCode, guiWallet.isoFiatCurrencyCode) : null,
-    primaryDisplayDenomination: guiWallet ? getDisplayDenominationFull(state, selectedCurrencyCode) : null,
-    primaryExchangeDenomination: guiWallet ? getExchangeDenomination(state, selectedCurrencyCode) : null
-  }
-}
+const selector = (state: RootState) => ({
+  selectedWalletId: state.ui.wallets.selectedWalletId,
+  selectedCurrencyCode: state.ui.wallets.selectedCurrencyCode
+})
 
 export default function ControlPanel() {
   console.debug('RENDER CONTROL PANEL')
 
   const [isViewUserList, setIsViewUserList] = useState(false)
 
-  const toggleUserList = () => setIsViewUserList(!isViewUserList)
+  const toggleUserList = (value: boolean) => setIsViewUserList(value)
 
   const dispatch = useDispatch()
 
@@ -56,34 +40,24 @@ export default function ControlPanel() {
   const styles = getStyles(theme)
 
   // Get props from redux state
-  const { selectedCurrencyCode, selectedWalletId, guiWallet, exchangeRate, primaryDisplayDenomination, primaryExchangeDenomination }: StateProps =
-    useSelector(selector)
+  const { selectedCurrencyCode, selectedWalletId }: StateProps = useSelector(selector, reduxShallowEqual)
 
-  if (!guiWallet) return null
+  if (!selectedWalletId) return null
 
   // // Create redux actions
   const onLogout = () => dispatch(logoutRequest())
   const onSelectWallet = (walletId: string, currencyCode: string) => dispatch(selectWalletFromModal(selectedWalletId, selectedCurrencyCode))
 
-  // // Create variables
-  const currencyLogo = getCurrencyIcon(guiWallet.currencyCode, selectedCurrencyCode).symbolImage
-  const secondaryExchangeDenomination = getDenomFromIsoCode(guiWallet.fiatCurrencyCode)
-
   return (
     <SceneWrapper hasHeader={false} hasTabs={false} isGapTop={false} background="none">
       <View style={styles.panel}>
-        {isViewUserList ? <View style={styles.disable} /> : null}
-        <PanelHeader
-          currencyLogo={currencyLogo}
-          exchangeRate={exchangeRate}
-          selectedCurrencyCode={selectedCurrencyCode}
-          primaryDisplayDenomination={primaryDisplayDenomination}
-          primaryExchangeDenomination={primaryExchangeDenomination}
-          fiatCurrencyCode={guiWallet.fiatCurrencyCode}
-          secondaryExchangeDenomination={secondaryExchangeDenomination}
-          toggleUserList={toggleUserList}
-        />
-        <PanelBody onSelectWallet={onSelectWallet} onLogout={onLogout} />
+        <PanelDisable isDisable={isViewUserList} />
+        <View style={styles.header}>
+          <PanelLogo />
+          <PanelCurrency />
+          <AccountSwitcher onSwitch={toggleUserList} />
+        </View>
+        <PanelList onSelectWallet={onSelectWallet} onLogout={onLogout} />
         <Separator style={styles.separator} />
       </View>
     </SceneWrapper>
@@ -94,20 +68,22 @@ const getStyles = cacheStyles((theme: Theme) => ({
   panel: {
     flex: 1,
     backgroundColor: theme.panelBackground,
-    position: 'relative'
+    position: 'relative',
+    paddingHorizontal: theme.rem(1),
+    paddingTop: theme.rem(9.5)
   },
   separator: {
     marginBottom: theme.rem(1.5),
     marginTop: theme.rem(1)
   },
-  disable: {
-    backgroundColor: '#87939E',
-    opacity: 0.8,
+  header: {
+    borderBottomRightRadius: theme.rem(2),
+    borderBottomLeftRadius: theme.rem(2),
+    paddingHorizontal: theme.rem(1),
     position: 'absolute',
-    zIndex: 1,
-    top: 0,
     left: 0,
     right: 0,
-    bottom: 0
+    backgroundColor: theme.panelBackground,
+    zIndex: 2
   }
 }))
