+++ src/modules/UI/components/ControlPanel/components/PanelList.js
